app/requirements.txt
Django==5.2
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-yasg==1.21.10
pillow==11.3.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10